Refactor TicTacToe game structure by modularizing code into separate files
- Moved game logic functions to logic.cpp and logic.h for better organization.
- Created board.cpp and board.h to handle board-related functions.
- Separated player-related functions into player.cpp and player.h.
- Isolated computer AI functions into computer.cpp and computer.h.
- Added display functions to display.cpp and display.h for rendering the game board and scores.
- Removed redundant function declarations from TicTacToe.cpp to streamline the main game file.
