+++ board.cpp
@@ -1,0 +1,6 @@
+#include "board.h"
+
+void clearBoard(string *spaces)
+{
+    fill(spaces, spaces + 9, " ");
+}
