+++ board.h
@@ -1,0 +1,13 @@
+#ifndef BOARD_H
+#define BOARD_H
+
+#include <string>
+using namespace std;
+
+/**
+ * Resets the game board by filling all spaces with empty strings
+ * @param spaces Array containing the board state
+ */
+void clearBoard(string *spaces);
+
+#endif
